+++ Sources/Foundation/Double.swift
@@ -23,3 +23,14 @@ extension Double: DoubleConvertible {
         return Double(self)
     }
 }
+
+public extension Double {
+    /// Creates a string version of this Double value without ".0" if this contains an integer number.
+    var withoutZeros: String {
+        let formatter = NumberFormatter()
+        let number = NSNumber(value: self)
+        formatter.minimumFractionDigits = 0
+        formatter.maximumFractionDigits = 16 //maximum digits in Double after dot (maximum precision)
+        return String(formatter.string(from: number) ?? "")
+    }
+}